split utility funcs file into search and mock
